pyproject.toml
[tool.poetry]
name = "training_tracker"
version = "0.1.0"
description = ""
authors = ["Maryia Afanasyeva <Maryia.Afanasyeva@itechart-group.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.2.1"
flask-marshmallow = "^0.14.0"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
python-dotenv = "^0.20.0"
psycopg2-binary = "^2.9.3"
click = "8.0.4"
PyJWT = "^2.4.0"
pytest = "^7.3.2"
decouple = "^0.0.7"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
isort = "^5.10.1"
pylint = "^2.14.5"
black = "^22.6.0"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
